+++ docs/changelog.md
@@ -20,6 +20,7 @@ nav_order: 2
 
 - Use manually configured IP secure tunnel password over loading it from keyring.
 - Add methods to Keyring to get interfaces by individual address (host or tunnel).
+- Remove `InterfaceWithUserIdNotFound` and `InvalidSignature` errors in favor of `InvalidSecureConfiguration`.
 
 ### Management